Changed the way the model keymap is stored. Keymap now stores only the present keys. The previous keymap is now a "Layout Keymap" and is still used to eliminate the unnecessary pin checks for the matrix scanning. The new keymap makes it easier to index into and get the scan code for a pressed key.
Added the implementation for storing pressed keys. The pressed keys are stored using usize bitmaps, which depend on the chip architecture and the number of keys.

Changed the keyboard scan codes from u8 to NonZeroU8.

Added a check for new reports coming from the host over HID. I need to see how to implement `SET_REPORT` and `GET_REPORT`.

Added an active keymap which gets set when the board starts. This should read the keymap stored in the EEPROM (TODO) or if one doesn't exist, get the default one from the device configuration.

Added new settings for the release build.

Other small changes.