Reworked the way keymaps and reports are being built and used.
The keyboard now has a `KeymapsState` and a `ReportState`. This makes it easier and safer to work with them. For keymaps a static mut value was removed and there is support now for active layers which can be enabled/disabled. For the report some extra info is stored to enable other functionality like switching between layers and differentiating between released and newly pressed keys.

To enable layer functionality some keycodes from HID are used. This is very likely to change and it's only used for now until I find a proper solution.

Other changes:

Removed the `KeyboardConfiguration` struct. The `Keymaps` struct is going to be used directly instead.
Added a `link_section` attribute for `MaybeUninit` `static mut` values.
Changed the `CountDown` in the time module to return a proper error instead of a &str.
Set the `flip-link` linker for all arm targets.
Updated the version of crate dependencies.
Raised rust toolchain nightly version.
Moved the version field to the workspace manifest file to make it easier to increment the version for all workspace crates.
Added more rust and and clippy lints.